renaming and update navbar
